Style: 도메인 공용 클래스 글로벌 패키지로 이동
