+++ src/main/java/org/guzzing/studayserver/domain/member/model/vo/NickName.java
@@ -1,5 +1,6 @@
 package org.guzzing.studayserver.domain.member.model.vo;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import java.util.Objects;
 import lombok.AccessLevel;
@@ -9,21 +10,27 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class NickName {
 
-    private String nickName;
+    public static final int NAME_MAX_LENGTH = 200;
 
-    public NickName(String nickName) {
-        validateName(nickName);
-        this.nickName = nickName;
+    @Column(name = "nickname")
+    private String value;
+
+    public NickName(String value) {
+        validateName(value);
+        this.value = value;
     }
 
     private void validateName(String name) {
         if (name == null || name.isBlank()) {
-            throw new IllegalArgumentException();
+            throw new IllegalArgumentException("닉네임은 빈 값일 수 없습니다.");
+        }
+        if (NAME_MAX_LENGTH < name.length()) {
+            throw new IllegalArgumentException(String.format("닉네임은 %d 글자를 넘을 수 없습니다.", NAME_MAX_LENGTH));
         }
     }
 
-    public String getNickName() {
-        return nickName;
+    public String getValue() {
+        return value;
     }
 
     @Override
@@ -35,12 +42,12 @@ public class NickName {
             return false;
         }
         NickName nickName1 = (NickName) o;
-        return Objects.equals(nickName, nickName1.nickName);
+        return Objects.equals(value, nickName1.value);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(nickName);
+        return Objects.hash(value);
     }
 
 }
