+++ src/main/java/org/guzzing/studayserver/domain/review/controller/ReviewRestController.java
@@ -5,7 +5,7 @@ import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 import jakarta.validation.Valid;
-import org.guzzing.studayserver.domain.auth.member_id.MemberId;
+import org.guzzing.studayserver.global.common.member.MemberId;
 import org.guzzing.studayserver.domain.review.controller.dto.request.ReviewPostRequest;
 import org.guzzing.studayserver.domain.review.controller.dto.response.ReviewPostResponse;
 import org.guzzing.studayserver.domain.review.controller.dto.response.ReviewableResponse;
