+++ src/test/java/org/guzzing/studayserver/domain/child/model/ChildNicknameTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.stream.Stream;
-import org.guzzing.studayserver.domain.member.model.NickName;
+import org.guzzing.studayserver.domain.member.model.vo.NickName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
